Refactor: keeping file names and variable names in consistence
