+++ client/actions/index.js
@@ -21,7 +21,7 @@ export const TYPES = {
   UPDATE_HEIGHT:                'UPDATE_HEIGHT',
   UPDATE_WEIGHT:                'UPDATE_WEIGHT',
   UPDATE_SOCIAL_SECURITY:       'UPDATE_SOCIAL_SECURITY',
-  UPDATE_VOTER_CITIZEN_STATUS:  'UPDATE_VOTER_CITIZEN_STATUS'
+  UPDATE_CITIZEN_STATUS:        'UPDATE_CITIZEN_STATUS'
 };
 
 export const updateLegalName            = generateAction(TYPES.UPDATE_LEGAL_NAME);
@@ -35,4 +35,4 @@ export const updateSex                  = generateAction(TYPES.UPDATE_SEX);
 export const updateHeight               = generateAction(TYPES.UPDATE_HEIGHT);
 export const updateWeight               = generateAction(TYPES.UPDATE_WEIGHT);
 export const updateSocialSecurity       = generateAction(TYPES.UPDATE_SOCIAL_SECURITY);
-export const updateVoterCitizenStatus   = generateAction(TYPES.UPDATE_VOTER_CITIZEN_STATUS);
+export const updateCitizenStatus        = generateAction(TYPES.UPDATE_CITIZEN_STATUS);
