+++ client/presentations/summary/index.js
@@ -11,7 +11,7 @@ import HairColor         from './hair-color.jsx';
 import Height            from './height.jsx';
 import Weight            from './weight.jsx';
 import SocialSecurity    from './social-security.jsx';
-import CitizenStatus     from './voter-citizen-status.jsx';
+import CitizenStatus     from './citizen-status.jsx';
 
 export {
   Empty,
